Add row count information and refresh listener
Added getRowCountRaw() and getRowCountFiltered() methods to return the
number of rows processed during the last refresh().
Added addRefreshListener() to allow clients to add a listener to be
notified when a refresh has been processed.

You can use these methods to update a label displaying the number of
filtered and total rows in the viewer that will automatically be updated
with any viewer refresh.

The counting must be enabled using setRowCountEnabled(true).